separate the tests to reporoot/tests
